requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-image>=0.19.0
colorspacious>=1.1.2
jsonschema>=4.0.0
click>=8.0.0
pandas>=1.3.0
